Initial Commit: Vipra-technical Frontend
This is the Frontend for Vipra technologies technical test. I have put this in a seperate Repo to make it easier to push and show my commits and then will merge the Django Backend with the frontend into a singular file.

Currently it can get all the contacts from the backend and has input boxes to show where the user will place the information for updating and adding contacts to the database.